+++ biopipen/scripts/scrna/MarkersFinder.R
@@ -417,8 +417,12 @@ process_allmarkers <- function(markers, object, comparison_by, plotcases, casena
         plotargs$markers <- markers
         plotargs$object <- object
         plotargs$comparison_by <- comparison_by
-        if (subset_by_group)
-            plotargs$subset_by <- groupname
+        if (
+            subset_by_group ||
+            plotargs$plot_type %in% c("jitter", "jitter_log2fc", "jitter_pct", "heatmap_log2fc", "heatmap_pct", "dot_log2fc", "dot_pct")
+        ) {
+            plotargs$subset_by <- plotargs$subset_by %||% groupname
+        }
         plotargs$outprefix <- file.path(info$prefix, slugify(plotname))
         do_call(VizDEGs, plotargs)
         reporter$add2(